# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Veil Panel API
  version: "0.6.3"
  description: |
    HTTP management API exposed by the Veil Panel (`veil serve`).

    ## Authentication model
    Veil is fail-closed for public Panel exposure. A non-loopback `--listen`
    address requires native TLS, API token authentication (`--auth-token` or
    `VEIL_API_TOKEN`), user/session authentication, and authenticated metrics.
    Caddy exposure requires user/session authentication and authenticated
    metrics even though the Veil listener remains on loopback.

    An unconfigured local loopback listener exposes only the first-run setup
    flow needed to create the initial administrator. Setup is unavailable
    through direct or Caddy exposure.

    `/api/*` routes accept either `X-Veil-Token`, `Authorization: Bearer`, or
    a `veil_session` cookie created by `/api/auth/login`. Mutating requests
    made with a cookie session must also send `X-CSRF-Token`, and only `admin`
    sessions may mutate management state. The sole viewer mutation exception
    is `POST /api/auth/locale`, which updates only the authenticated user's own
    language preference and rejects static API-token callers.

    Browser sessions survive Panel restarts. Only hashes of the session cookie
    and CSRF token are persisted. Sessions expire after 30 minutes idle or 24
    hours absolute and are revoked when a user's password, role, or existence
    changes.

...
    until a valid session exists.

...
  license:
...
servers:
...
security:
  - veilToken: []
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /api/setup/complete:
    post:
      x-veil-role: public
      tags: [setup]
      summary: Create the initial administrator
      description: Single-use operation available only on an unconfigured local loopback listener.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupCompleteRequest"
            examples:
              initial-admin:
                value:
                  username: admin
                  password: correct-horse-battery-staple
                  backupAcknowledged: true
                  locale: ru
      responses:
        "201":
          description: Initial administrator created and setup marked complete.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupCompleteResponse"
              examples:
                complete:
                  value:
                    completed: true
                    username: admin
                    role: admin
                    locale: ru
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Setup was already completed or a user already exists.
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorText"
...
  /api/auth/locale:
    post:
      x-veil-role: self-service
      tags: [auth]
      summary: Update the current browser user's locale
      description: |
        Self-service exception available to both `admin` and `viewer` browser
        sessions. Requires the `veil_session` cookie and `X-CSRF-Token`,
        updates only that session's username, and sets a one-year
        `veil_locale` preference cookie. Static API tokens cannot call this
        endpoint.
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocaleUpdateRequest"
            examples:
              russian:
                value:
                  locale: ru
      responses:
        "200":
          description: Locale persisted for the current user.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: SameSite `veil_locale` preference cookie.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocaleResponse"
              examples:
                russian:
                  value:
                    locale: ru
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/client-links/qr:
    post:
      x-veil-role: admin-secret
      tags: [client-links]
      summary: Render a client URI as a local QR PNG
      description: Generates the QR image inside Veil so client URIs are not sent to a third-party QR service.
      security:
        - veilToken: []
        - bearerAuth: []
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientLinkQRRequest"
            examples:
              hysteria2:
                value:
                  uri: hysteria2://example-secret@example.com:443/?sni=example.com#veil
      responses:
        "200":
          description: QR image for the submitted client URI.
          content:
            image/png:
              schema:
                type: string
                contentEncoding: base64
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: Submitted URI is too large for QR export.
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorText"
...
  /api/apply/plan:
    post:
      x-veil-role: viewer
      tags: [apply]
      summary: Preview the apply plan without writing live files
      security:
        - veilToken: []
        - bearerAuth: []
        - sessionCookie: []
          csrfToken: []
      responses:
        "200":
          description: Apply plan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyPlanResponse"
              examples:
                restart-preview:
                  value:
                    valid: true
                    errors: []
                    configs:
                      - /etc/veil/generated/mieru/server_config.json
                    actions:
                      - validate management state
                      - stage generated configs
                      - restart veil-mieru.service
                    runtimes:
                      - veil-mieru.service
                    issues: []
                    operations:
                      - type: promote_file
                        source: /etc/veil/generated/mieru/server_config.json
                        destination: /etc/veil/live/mieru/server_config.json
                        interruptionRisk: reload
                        rollbackAvailable: true
                        validationSource: render-and-live-host
                      - type: restart_service
                        unit: veil-mieru.service
                        interruptionRisk: connection-drop
                        rollbackAvailable: true
                        validationSource: managed-unit-catalog
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
  /api/profiles/ru-recommended/preview:
    post:
      x-veil-role: viewer
      tags: [apply]
      summary: Preview the ru-recommended install profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RURecommendedPreviewRequest"
            examples:
              caddy:
                value:
                  domain: vpn.example.com
                  email: admin@example.com
                  panelAccess: caddy
      responses:
        "200":
          description: Profile preview.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RURecommendedPreviewResponse"
              examples:
                caddy:
                  value:
                    domain: vpn.example.com
                    email: admin@example.com
                    panelAccess: caddy
                    panelUrl: https://vpn.example.com/4f10dbe4/
                    caddyfile: |
                      vpn.example.com {
                        reverse_proxy 127.0.0.1:2096
                      }
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/tools/dns-lookup:
    post:
      x-veil-role: viewer
      tags: [diagnostics]
      summary: DNS lookup diagnostic
      security:
        - veilToken: []
        - bearerAuth: []
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DNSLookupRequest"
      responses:
        "200":
          description: DNS lookup result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DNSLookupResult"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/tools/speedtest:
    post:
      x-veil-role: viewer
      tags: [diagnostics]
      summary: Speedtest diagnostic
      security:
        - veilToken: []
        - bearerAuth: []
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmptyObject"
      responses:
        "200":
          description: Speedtest result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeedtestResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /api/v1/traffic/summary:
    get:
      x-veil-role: viewer
      tags: [traffic]
      summary: Aggregate traffic totals + honest telemetry provider state
      security:
        - sessionCookie: []
        - bearerAuth: []
        - veilToken: []
      responses:
        "200":
          description: '{"state,providerCount,uploadBytes,downloadBytes,usedBytes}'
  /api/v1/traffic/top:
    get:
      x-veil-role: viewer
      tags: [traffic]
      summary: Top talkers ranked by cumulative usage
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Clients ranked by total bytes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficTopResponse"
  /api/v1/traffic/{id}:
    get:
      x-veil-role: viewer
      tags: [traffic]
      summary: Per-client cumulative traffic totals
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200":
          description: Upload/download totals with quota and depleted flag.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficTotals"
        "404":
          description: Client not found.
...
  /api/v1/traffic/stream:
    get:
      x-veil-role: viewer
      tags: [traffic]
      summary: Server-sent stream of live traffic snapshots
      description: Emits a JSON snapshot per client every ~5s as SSE data frames.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: text/event-stream of per-client snapshots.
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  securitySchemes:
    veilToken:
      type: apiKey
      in: header
      name: X-Veil-Token
    bearerAuth:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: veil_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
...
    ClientId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  responses:
    BadRequest:
      description: Request body, query string, or path failed validation.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
          examples:
            invalid:
              value: "invalid JSON\n"
    Unauthorized:
      description: Missing or invalid token/session.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Present on authenticated API challenges.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
          examples:
            missing:
              value: "missing or invalid API token or session\n"
    Forbidden:
      description: Authenticated principal is not allowed to perform the operation.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
          examples:
            csrf:
              value: "invalid or missing CSRF token\n"
    NotFound:
      description: Resource not found.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
          examples:
            not-found:
              value: "404 page not found\n"
    ServiceUnavailable:
      description: Runtime dependency is unavailable.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
    ValidationFailed:
      description: Candidate configuration failed authoritative validation and was not mutated.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationFailure"
          examples:
            busy-port:
              value:
                error:
                  code: validation_failed
                  message: configuration failed live validation
                issues:
                  - code: port_in_use
                    severity: error
                    field: port
                    inboundId: edge
                    message: TCP port 443 is already in use
                    remediation: Stop the conflicting service or choose another port.
                    source: live-host
...
  schemas:
    EmptyObject:
      type: object
      additionalProperties: false
    ErrorText:
      type: string
      description: Plain text error emitted by `writeError`; usually a single message line ending in `\n`.
    ErrorObject:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code:
          type: string
        message:
          type: string
...
    SetupCompleteRequest:
      type: object
      required: [username, password, backupAcknowledged]
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[\p{L}\p{N}._-]+$'
        password:
          type: string
          format: password
          minLength: 12
        backupAcknowledged:
          type: boolean
          const: true
        locale:
          $ref: "#/components/schemas/Locale"
    SetupCompleteResponse:
      type: object
      required: [completed, username, role, locale]
      properties:
        completed:
          type: boolean
          const: true
        username:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        locale:
          $ref: "#/components/schemas/Locale"
...
    Locale:
      type: string
      enum: [en, ru]
      description: Persisted Panel display language.
    LocaleUpdateRequest:
      type: object
      required: [locale]
      additionalProperties: false
      properties:
        locale:
          $ref: "#/components/schemas/Locale"
    LocaleResponse:
      type: object
      required: [locale]
      additionalProperties: false
      properties:
        locale:
          $ref: "#/components/schemas/Locale"
...
    UserRole:
      type: string
      enum: [admin, viewer]
...
    ClientLinkQRRequest:
      type: object
      required: [uri]
      additionalProperties: false
      properties:
        uri:
          type: string
          minLength: 1
          maxLength: 4096
          description: Client connection URI or config export string to encode as a QR image.
...
    ValidationIssue:
      type: object
      required: [code, severity, message, source]
      additionalProperties: false
      properties:
        code:
          type: string
          examples: [port_in_use]
        severity:
          type: string
          enum: [error, warning, info]
        field:
          type: string
          examples: [port]
        inboundId:
          type: string
          examples: [edge]
        message:
          type: string
        remediation:
          type: string
        source:
          type: string
          enum: [candidate, live-host, render, managed-unit-catalog]
...
    ValidationFailure:
      type: object
      required: [error, issues]
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/ErrorObject"
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
    ApplyOperation:
      type: object
      required: [type, interruptionRisk, rollbackAvailable, validationSource]
      additionalProperties: false
      properties:
        type:
          type: string
          enum: [promote_file, remove_file, reload_service, restart_service, disable_service, update_firewall]
        source:
          type: string
        destination:
          type: string
        unit:
          type: string
        interruptionRisk:
          type: string
          enum: [none, reload, connection-drop, lockout-risk]
        rollbackAvailable:
          type: boolean
        validationSource:
          type: string
    ApplyPlanResponse:
      type: object
      required: [valid, configs, actions, issues, operations]
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
        configs:
          type: array
          items:
            type: string
        actions:
          type: array
          items:
            type: string
        runtimes:
          type: array
          items:
            type: string
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
        operations:
          type: array
          items:
            $ref: "#/components/schemas/ApplyOperation"
...
    RURecommendedPreviewRequest:
      type: object
      required: [domain, email]
      additionalProperties: false
      properties:
        domain:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        panelAccess:
          type: string
          enum: [local, direct, caddy]
    RURecommendedPreviewResponse:
      type: object
      required: [domain, email, panelAccess]
      additionalProperties: false
      properties:
        domain:
          type: string
        email:
          type: string
          format: email
        panelAccess:
          type: string
          enum: [local, direct, caddy]
        panelUrl:
          type: string
          format: uri
        caddyfile:
          type: string
...
    DNSLookupRequest:
      type: object
      required: [hostname]
      additionalProperties: false
      properties:
        hostname:
          type: string
          examples: [example.com]
    DNSLookupResult:
      type: object
      required: [hostname, addresses]
      properties:
        hostname:
          type: string
        addresses:
          type: array
          items:
            type: string
        cname:
          type: string
        error:
          type: string
...
    SpeedtestResult:
      type: object
      required: [pingMs, downloadMbps, uploadMbps]
      properties:
        server:
          type: string
        pingMs:
          type: number
        downloadMbps:
          type: number
        uploadMbps:
          type: number
        raw:
          type: string
...
    TrafficTotals:
      type: object
      required: [clientId, uploadBytes, downloadBytes, totalBytes, depleted]
      properties:
        clientId:
          type: string
        uploadBytes:
          type: integer
          format: int64
        downloadBytes:
          type: integer
          format: int64
        totalBytes:
          type: integer
          format: int64
        quotaBytes:
          type: integer
          format: int64
          minimum: 0
          # Capped at Number.MAX_SAFE_INTEGER: the SPA consumes this as a
          # JS number; larger int64s would lose precision silently.
          maximum: 9007199254740991
        remainingBytes:
          type: integer
          format: int64
        depleted:
          type: boolean
...
    TrafficTopEntry:
      type: object
      required: [clientId, name, totalBytes]
      properties:
        clientId:
          type: string
        name:
          type: string
        uploadBytes:
          type: integer
          format: int64
        downloadBytes:
          type: integer
          format: int64
        totalBytes:
          type: integer
          format: int64
    TrafficTopResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TrafficTopEntry"